:art: separate MainPage into presentational and container components
